feat(svi/wandb): push gradient estimators to W&B
Add the optax hook[^1] that allows to store either all or the latest
estimate (mean or norm) of a gradient (notably, it will aggregate large
axis but preserve small ones, < 5 elements).
The values from the store, where the hook deposits the gradient
aggregated value, are then merged with the rest of the payload to
wandb.log.